Login button, Session storage, Login form added, all fully functional
